autotuner can handle weight in stopMeas now
git-svn-id: https://svn.ipd.kit.edu/repos/karcher/karcher/markus/Tuner@2745 1cde74c2-b0a7-11dd-9e77-cbfca2305f68

+++ Client/tuner.h
@@ -23,7 +23,7 @@ class Tuner {
 	int tRegisterSectionParameter(int sectionId, int *parameter);
 	int tRequestStart(int sectionId);
 	int tStop(int sectionId);
-	int tStop(int sectionId, int weight);
+	int tStop(int sectionId, double weight);
 	int tFinishTuning();
 	int tReset();
 	int tSetpersistence(int pers);
